Create a product using JDBC
It seems like you want to create a Product Management System using BDS (Backend Database System), Forms, and POJO (Plain Old Java Object) to manage products in a database. In this setup, we will use JDBC to interact with the database, Forms (such as Swing or AWT) for the user interface, and POJO for data representation.